% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refstudio-package.R
\docType{package}
\name{refstudio-package}
\alias{refstudio}
\alias{refstudio-package}
\title{refstudio: Tools to Extract and Process Bibliographical References}
\description{
A complete toolkit to extract and process bibliographical references.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://giperbio.github.io/refstudio/}
  \item \url{https://github.com/giperbio/refstudio/}
  \item Report bugs at \url{https://github.com/giperbio/refstudio/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
